+++ src-tauri/tests/manual_windows_live_smoke.rs
@@ -1,6 +1,6 @@
 /**
  * [INPUT]: 依赖显式 disposable clone/evidence 环境变量、共享 Windows 路径守卫、apply_language_inner、RealCommandRunner、runtime marker 与 PowerShell PID 窗口证据 helper
- * [OUTPUT]: 对外提供 ignored Windows live-clone 冒烟：可选单语或默认三语 QPA 原生启动、隔离 AppData、三类带完整 source 位图的自动 PNG、带零位图基线与严格计数增量的人工 Cog Pitch PNG、PID 清理及显式 English 恢复
+ * [OUTPUT]: 对外提供 ignored Windows live-clone 冒烟：可选单语或默认三语 QPA 原生启动、隔离 AppData、三类带完整 64 位 source 位图的自动 PNG、带零位图基线与严格计数增量的人工 Cog Pitch PNG、PID 清理及显式 English 恢复
  * [POS]: src-tauri/tests 的 Windows GUI 现场证据门；自动路径经有界 exact-HWND 前台门投递 A 键并锁定 EditShape 动作与长操作前缀，Cog Pitch 仅在 opt-in 后记录前后诊断并等待用户操作 sentinel clone，不创建场景、不运行脚本、不依赖 Qt UIA
  * [PROTOCOL]: 变更时更新此头部，然后检查 CLAUDE.md
  */
@@ -80,8 +80,8 @@ mod windows_live_smoke {
         original_fallback: u64,
         no_translation: u64,
         renderer_failure: u64,
-        translated_source_mask: u32,
-        fallback_source_mask: u32,
+        translated_source_mask: u64,
+        fallback_source_mask: u64,
     }
 
     #[derive(Debug, Deserialize)]
